split gundeck API
